requirements.local.txt
# Core dependencies
neo4j>=5.14.0
python-dotenv>=1.0.0
requests>=2.31.0

# Local development dependencies
openai>=1.0.0
sentence-transformers>=2.2.0
spacy>=3.7.0
numpy>=1.24.0
pandas>=2.0.0

# Web framework for local API
fastapi>=0.104.0
uvicorn>=0.24.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development tools
black>=23.0.0
flake8>=6.0.0
